pyproject.toml
[project]
name = "news-summary"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "feedparser==6.0.11",
    "anthropic>=0.50.0",
    "python-dotenv==1.0.0",
    "json-repair>=0.30.0",
]
